Updates to recipes and unit tests.
Added logic to state.py to handle legacy recipes.
