+++ data/recipes/aws_turbinia_ts.json
@@ -63,11 +63,13 @@
       "wants": ["GCEDiskFromImage"],
       "name": "TurbiniaGCPProcessor",
       "args": {
-        "turbinia_config_file": null,
         "disk_names": null,
         "project": "@gcp_project",
         "turbinia_recipe": "@turbinia_recipe",
         "turbinia_zone": "@turbinia_zone",
+        "turbinia_auth": "@turbinia_auth",
+        "turbinia_api": "@turbinia_api",
+        "incident_id": "@incident_id",
         "sketch_id": "@sketch_id"
       }
     }, {
@@ -96,6 +98,8 @@
     ["--token_password", "Optional custom password to decrypt Timesketch credential file with.", ""],
     ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
     ["--turbinia_zone", "Zone Turbinia is located in", "us-central1-f", {"format": "gcp_zone"}],
+    ["--turbinia_auth", "Flag to indicate whether Turbinia API server requires authentication.", false],
+    ["--turbinia_api", "Turbinia API server endpoint.", "http://127.0.0.1:8000"],
     ["--wait_for_timelines", "Whether to wait for Timesketch to finish processing all timelines.", true]
   ]
 }
