+++ data/recipes/gcp_turbinia_ts.json
@@ -11,7 +11,7 @@
     }],
     "modules": [{
         "wants": [],
-        "name": "TurbiniaAPIGCPProcessor",
+        "name": "TurbiniaGCPProcessor",
         "args": {
             "disk_names": "@disk_names",
             "project": "@analysis_project_name",
@@ -23,7 +23,7 @@
             "sketch_id": "@sketch_id"
         }
     }, {
-        "wants": ["TurbiniaAPIGCPProcessor"],
+        "wants": ["TurbiniaGCPProcessor"],
         "name": "TimesketchExporter",
         "args": {
             "incident_id": "@incident_id",
