+++ data/recipes/upload_turbinia_legacy.json
@@ -1,5 +1,5 @@
 {
-    "name": "upload_turbinia",
+    "name": "upload_turbinia_legacy",
     "short_description": "Uploads arbitrary files to Turbinia and downloads results.",
     "description": "Uploads arbitrary files to Turbinia for processing. The recipe will wait for Turbinia to return with results and will download them back to the filesystem. The Turbinia system needs to be accessible via SSH.",
     "modules": [{
@@ -32,19 +32,17 @@
           }
     }, {
         "wants": ["SCP-Upload"],
-        "name": "TurbiniaArtifactProcessor",
+        "name": "TurbiniaArtifactProcessorLegacy",
         "args": {
+            "turbinia_config_file": "@turbinia_config",
             "project": null,
-            "turbinia_auth": "@turbinia_auth",
             "turbinia_recipe": "@turbinia_recipe",
             "turbinia_zone": "@turbinia_zone",
-            "turbinia_api": "@turbinia_api",
-            "incident_id": "@incident_id",
             "output_directory": null,
             "sketch_id": null
         }
     }, {
-          "wants": ["TurbiniaArtifactProcessor"],
+          "wants": ["TurbiniaArtifactProcessorLegacy"],
           "runtime_name": "SCP-Download",
           "name": "SCPExporter",
           "args": {
@@ -65,8 +63,7 @@
         ["--destination_turbinia_dir", "Destination path in Turbinia host to write the files to.", null],
         ["--hostname", "Remote host.", null, {"format": "hostname"}],
         ["--directory", "Directory in which to copy and compress files.", null],
-        ["--turbinia_auth", "Flag to indicate whether Turbinia API server requires authentication.", false],
-        ["--turbinia_api", "Turbinia API server endpoint.", "http://127.0.0.1:8000"],
+        ["--turbinia_config", "Turbinia config file to use.", null],
         ["--local_turbinia_results", "Directory where Turbinia results will be downloaded to.", null],
         ["--turbinia_zone", "The GCP zone the disk to process and Turbinia workers are in.", "us-central1-f", {"format": "gcp_zone"}],
         ["--sketch_id", "Timesketch sketch ID.", null, {"format": "regex", "regex": "^\\d+$"}]
